Use lightuserdata for entity IDs in Lua scripting
Entity IDs previously crossed the Lua scripting boundary as (possibly
truncated) integers. They now cross as Lua lightuserdata values, which
preserve the full 64-bit entity ID. The generated ulong marshaller pushes
and reads lightuserdata, and the uint/ushort scalar marshallers are
re-routed to the long (integer) marshaller so that colors, quantities, and
calendar times keep integer semantics.

All hand-written scripting entry points that read or write entity IDs now
require lightuserdata with a strict type check: the Entities, Data, and
Scripting libraries, the generated component bindings, and the callback
invocation paths (entity callbacks, entity script callbacks, OnInteract
arguments, and chat command callbacks). The Entities module gains new
helper functions FormatEntityId, ToEntityId, and ToTemplateEntityId, its
entity ID constants are pushed as lightuserdata, and IsTemplate becomes a
C# library function since Lua 5.4 forbids ordered comparisons on
lightuserdata. System time remains a Lua number: Time.GetSystemTime and
Time.FutureSystemTime now return double, and Kinematics.StopSystemTime
changes from ulong to double; the field is not persisted. To support this,
System.Double was added to the Lua library generator's predefined
marshaller types. The related stdlib scripts and Lua test suites were
updated accordingly.

Two transient key-value formats change as a side effect: the
Sovereign.TimedChange scheduling key now uses the formatted hex template
ID, and the Sovereign.ItemSpawn.ItemId entity-data value is now stored as
a formatted hex string (read back with base-16 tonumber) since
lightuserdata values cannot be stored in the entity key-value store; both
only affect in-flight scheduling/spawn state and reset harmlessly.

Tested: full clean rebuild of Sovereign.sln (Debug) with 0 warnings and 0
errors; dotnet test (Debug) with 202/202 tests passing across
TestEngineUtil, TestEngineCore, TestPersistence, and TestServerCore; Lua
script test harness (util/test-scripts.sh) with all 15 suites / 63 tests
passing against a fresh migration-seeded database; run-server skill
integration run with clean startup, all standard scripts loading without
Lua errors, and an ad-hoc probe (kept as Data/Scripts/Test/
LightUserDataProbe.lua.txt) validating FormatEntityId/ToEntityId/
ToTemplateEntityId/IsTemplate and a scheduled stop driven by
Time.GetSystemTime; docs manual rebuilt with sphinx (venv absent in this
environment, so requirements.txt was installed via pip) with no new
warnings.

+++ src/Common/EngineCore/Components/Types/Kinematics.cs
@@ -38,5 +38,5 @@ public struct Kinematics
     /// <summary>
     ///     If non-zero, specifies the system time at which motion should stop.
     /// </summary>
-    [ScriptableField] public ulong StopSystemTime;
+    [ScriptableField] public double StopSystemTime;
 }
